requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
requests>=2.28.0
python-dotenv>=0.19.0
openpyxl>=3.0.0
pytest>=7.0.0
jupyter>=1.0.0